Added an approach to identifying and handling file with arrays of objects
The first element of the file should be the track key for the rest of the elements. No file name will include those elements.

So, because files are being used, a file name will have a key identifying the object _tracking number. The number will also be in the object. For files that are lists of objects, the name of the file will include the tracking num with a symbol identifying a list. A similar marker, defined by the application will be the first element of the array.